05requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
psycopg==3.1.18
python-dotenv==1.0.0
cloudinary==1.36.0
pytz==2023.3
chart.js==0.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0